Add font-awesome codepoints to the dmenuunicode menu
This adds the codepoints for font-awesome to ~/.fontawesome in the same
format as ~/.emojii and modifies the grep search pattern in dmenuunicode
to search both ~/.font-awesome and ~/.emojii including both in the list
of values sent to dmenu.

Removing the ~/.fontawesome file will revert to the previous behaviour.

This requires the fontawesome-fonts package to be installed to get the
additional codepoints.